requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.1.2
python-multipart
sse-starlette
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
chromadb
sentence-transformers
pypdf
python-docx
beautifulsoup4
unstructured
python-dotenv
duckduckgo-search
pytest
httpx